Automate AI battle
User will call attacks on its own board over and over to simulate the
AI opponent attacking. This solution was way easier than making a bunch
of changes to have two "real" AI battle it out. The only purpose
is to test how good the AI is anyways so this does not matter.